@@ -17,26 +17,26 @@
   },
   "homepage": "https://github.com/oeran93/food305#readme",
   "dependencies": {
-    "aws-sdk": "^2.79.0",
+    "aws-sdk": "^2.92.0",
     "babel-core": "^6.10.4",
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
-    "body-parser": "^1.15.2",
+    "body-parser": "^1.17.2",
     "client-sessions": "^0.7.0",
     "crypto": "0.0.3",
-    "express": "^4.14.0",
+    "express": "^4.15.3",
     "jquery": "^3.1.1",
-    "json-loader": "^0.5.4",
+    "json-loader": "^0.5.7",
     "moment": "^2.15.2",
     "moment-timezone": "^0.5.13",
-    "mongoose": "^4.6.2",
-    "prop-types": "^15.5.8",
-    "react": "^15.3.2",
+    "mongoose": "^4.11.4",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
     "react-bootstrap": "^0.30.6",
-    "react-dom": "^15.3.2",
-    "react-router-dom": "^4.1.1",
-    "stripe": "^4.23.0",
+    "react-dom": "^15.6.1",
+    "react-router-dom": "^4.1.2",
+    "stripe": "^4.23.1",
     "twilio": "^2.11.1",
     "underscore": "^1.8.3",
     "webpack": "^1.13.2"
