@@ -3,7 +3,7 @@ const env = process.env
 
 module.exports = function () {
 	const auth = `mongodb://${env.VIMI_DB_USERNAME}:${env.VIMI_DB_PASSWORD}@localhost/${env.VIMI_DB_NAME}`
-	mongoose.connect(auth,
+	mongoose.connect(auth, {useMongoClient: true},
 		(err, res) => {
 			if (err) {
 				console.log('db error: ', err.message);
