@@ -8,6 +8,8 @@ module.exports = {
 
   credit_card_percentage: .03,
 
+  trial_days: 30,
+
   app_name: "Vimi",
 
   twilio_accountsid: env.VIMI_TWILIO_ACCOUNTSID,
@@ -20,6 +22,8 @@ module.exports = {
 
   order_date_format: "MM-DD-YYYY HH:mm",
 
+  iso_date_format: "YYYY-MM-DDTHH:mm:ssZ",
+
   day_date_format: "MM-DD-YYYY",
 
   info_email: "info@vimifood.com",
@@ -30,6 +34,10 @@ module.exports = {
 
   curator_port: env.CURATOR_PORT,
 
-  test_port: env.TEST_PORT
+  test_port: env.TEST_PORT,
+
+  ok_financial_statuses: ['active', 'trialing'],
+
+  failed_financial_statuses: ['past_due']
 
 }
