@@ -2,6 +2,7 @@ const React        = require('react')
 const errors       = require('../../../../tools/errors.js')
 const PropTypes    = require('prop-types')
 const ajx = require('../../../../tools/ajax.js')()
+const confirmation = require('../../../../tools/confirmation.js')()
 
 class Basic_Info extends React.Component {
 
@@ -11,7 +12,8 @@ class Basic_Info extends React.Component {
       phone: "",
       name: "",
       email: "",
-      station: ""
+      station: "",
+      terms_and_conditions: false
     }
   }
 
@@ -24,28 +26,39 @@ class Basic_Info extends React.Component {
       }
     })
   }
+  
+  terms_and_conditions () {
+    this.setState((prevState) => ({
+      terms_and_conditions: !prevState.terms_and_conditions
+    }))
+  }
 
   handle_change (event) {
     let state = this.state
     state[event.target.id] = event.target.value
+    console.log(state)
     this.setState(state)
   }
 
   send_info () {
     let {change_step} = this.props
-    ajx.call({
-      method: 'POST',
-      url: '/create_user',
-      data : this.state,
-      success: (res) => change_step({step: 1, phone: this.state.phone}),
-      error: (res) => {if (res.error.number == errors.user_exists.number) change_step({step: 3})},
-      show_messages: true,
-      show_loading: true
-    })
+    let {terms_and_conditions, phone} = this.state
+    if (terms_and_conditions) {
+      ajx.call({
+        method: 'POST',
+        url: '/create_user',
+        data : this.state,
+        success: (res) => change_step({step: 1, phone}),
+        error: (res) => {if (res.error.number == errors.user_exists.number) change_step({step: 3})},
+        show_messages: true,
+        show_loading: true
+      })
+    } else confirmation.failure("Please accept the terms and conditions")
+    
   }
 
   render () {
-    let {phone, name, email} = this.state
+    let {phone, name, email, terms_and_conditions} = this.state
     let {autofocus} = this.props
     return (
       <div className='row access'>
@@ -85,6 +98,9 @@ class Basic_Info extends React.Component {
             onChange={this.handle_change.bind(this)}
             onKeyPress={(t) => {if (t.charCode === 13) this.send_info.bind(this)()}}
           />
+        </div>
+        <div className="col-xs-12 input">
+          <h5> <input type="checkbox" onChange={this.terms_and_conditions.bind(this)} checked={terms_and_conditions}/> <a href="/terms_and_conditions.pdf"> I Accept the terms and conditions</a></h5>
         </div>
         <div className='col-xs-12'>
           <button className='btn red-btn pull-right' onClick={this.send_info.bind(this)}>
