@@ -2,36 +2,57 @@ const React          = require('react')
 const format         = require('../../../tools/format.js')
 const generics        = require('../../../tools/generics.js')
 const PropTypes      = require('prop-types')
+import {Modal} from 'react-bootstrap'
+import Meal_Payment from './meal_payment.jsx'
 
-const Meal = function(props) {
-  let {meal, toggleModal} = props
-  let {_id, name, price, image, tags, orders, description} = meal
-  return (
-    <div name={_id} className='col-sm-6 col-md-4 col-lg-3 meal'>
-      <div className='thumbnail clearfix'>
-        {orders.length > 0 && <span className='meal-orders'> {orders.length} </span>}
-        <img src={'images/meals/' + image} alt={name} className='meal-picture'/>
-        <div className='clearfix'>
-          <h5 title={name} className='meal-name text-uppercase text-center'>
-            {name}
-          </h5>
-          <p>{description}</p>
-          <button
-              className="btn red-btn meal-btn pull-left"
-              onClick={() => toggleModal('payment_modal', {open: true, meal: meal, price: price})}
-          >
-            Buy
-          </button>
-          <span className='green-label price pull-right'> {'$ '+price} </span>
+class Meal extends React.Component {
+  
+  constructor (props) {
+    super(props)
+    this.state = {
+      modal: {open: false}
+    }
+  }
+  
+  toggle_modal () {
+    this.setState((old_state) => ({
+      modal: {open: !old_state.modal.open}
+    }))
+  }
+  
+  render () {
+    let {meal} = this.props
+    let {_id, name, price, image, tags, orders, description} = meal
+    let {modal} = this.state
+    return (
+      <div name={_id} className='col-sm-6 col-md-4 col-lg-3 meal'>
+        <div className='thumbnail clearfix'>
+          {orders.length > 0 && <span className='meal-orders'> {orders.length} </span>}
+          <span className='label label-default meal-price'> {'$ '+price} </span>
+          <img src={'images/meals/' + image} alt={name} className='meal-picture'/>
+          <div className='clearfix'>
+            <h5 title={name} className='meal-name text-center'>
+              {name}
+            </h5>
+            <p className="meal-description">{description}</p>
+            <button className="btn red-btn meal-btn pull-right" onClick={this.toggle_modal.bind(this)}>
+              Buy
+            </button>
+          </div>
         </div>
+        <Modal show={modal.open} onHide={this.toggle_modal.bind(this)}>
+          <Modal.Header closeButton></Modal.Header>
+          <Modal.Body>
+            <Meal_Payment autofocus={true} meal={meal}/>
+          </Modal.Body>
+        </Modal>
       </div>
-    </div>
-  )
+    )
+  }
 }
 
 Meal.propTypes = {
-  meal: PropTypes.object.isRequired,
-  toggleModal: PropTypes.func
+  meal: PropTypes.object.isRequired
 }
 
 module.exports = Meal
