@@ -1,13 +1,15 @@
 import React from 'react'
 import _ from 'underscore'
-import Home from './home.jsx'
+import Menu from './menu.jsx'
 import Nav_Bar from './nav_bar.jsx'
 import Access from './access/access.jsx'
 import About from './about.jsx'
-import Payment from './payment/payment.jsx'
+import Subscribe from './subscribe.jsx'
+import Profile from './profile.jsx'
+import Failed_Billing from './failed_billing.jsx'
 import Footer from './footer.jsx'
 import {Modal} from 'react-bootstrap'
-import {BrowserRouter as Router, Route} from 'react-router-dom'
+import {HashRouter as Router, Route,} from 'react-router-dom'
 const ajx = require('../../../tools/ajax.js')()
 
 class Root extends React.Component {
@@ -16,13 +18,12 @@ class Root extends React.Component {
     super(props)
     this.state = {
       render: false,
-      user: null,
-      access_modal: {open: false, step: 3},
-      payment_modal: {open: false, amount: "0", meal: {}}
+      access_modal: {open: false, step: 3}
     }
+    window.store.setAppRoot(this)
   }
 
-  toggleModal (modal, info) {
+  toggle_modal (modal, info) {
     let state = this.state
     state[modal] = _.extend(state[modal],info)
     this.setState(state)
@@ -33,55 +34,57 @@ class Root extends React.Component {
       method: 'GET',
       url: '/profile',
       success: data => {
-        this.setState({
-          user: data.user,
-          render: true
-        })
+        window.store.set('user', data.user)
+        this.setState({render: true})
       }
     })
   }
 
   render () {
-    let {render, user, access_modal, payment_modal} = this.state
+    const {render, access_modal} = this.state
+    const user = window.store.get('user')
     if (!render) return null
-    else {
-      return (
-        <Router>
-          <div>
-            {/*Navigation Bar*/}
-            <Nav_Bar user={user} toggleModal={this.toggleModal.bind(this)}/>
-            {/*Main Content*/}
-            <Route exact path='/' component={
-                () => user ? <Home user={user} toggleModal={this.toggleModal.bind(this)}/> : <About toggleModal={this.toggleModal.bind(this)}/>
-              }
-            />
-            {/*Access Modal*/}
-            <Modal show={access_modal.open} onHide={() => this.toggleModal.bind(this)('access_modal', {open:false})}>
-              <Modal.Header closeButton></Modal.Header>
-              <Modal.Body>
-                <Access step={access_modal.step} />
-              </Modal.Body>
-            </Modal>
-            {/*Payment Modal*/}
-            <Modal show={payment_modal.open} onHide={() => this.toggleModal.bind(this)('payment_modal', {open:false})}>
-              <Modal.Header closeButton></Modal.Header>
-              <Modal.Body>
-                <Payment autofocus={true} price={payment_modal.price} meal={payment_modal.meal} user={user}/>
-              </Modal.Body>
-            </Modal>
-            {/*Loading*/}
-            <div className="loading-background">
-              <div className="loading">
-                <div className="lds-ripple">
-                  <div></div>
-                  <div></div>
-                </div>
+    return (
+      <Router>
+        <div>
+          <div className="confirmation-wrapper">
+            <div className='confirmation success'></div>
+            <div className='confirmation failure'></div>
+          </div>
+          <Nav_Bar toggleModal={this.toggle_modal.bind(this)}/>
+          {user 
+            ?
+              <div>
+                <Route exact path='/' component={() => <Menu/>}/>
+                <Route path="/subscribe" component={Subscribe} />
+                <Route path="/failed_billing" component={Failed_Billing} />
+                <Route path="/profile" component={() => <Profile/>} />
+              </div>
+            :
+              <div>
+                <Route exact path='/' component={() => <About toggleModal={this.toggle_modal.bind(this)}/>}/>
+              </div>
+          }
+          {/*Access Modal*/}
+          <Modal show={access_modal.open} onHide={() => this.toggle_modal.bind(this)('access_modal', {open:false})}>
+            <Modal.Header closeButton></Modal.Header>
+            <Modal.Body>
+              <Access step={access_modal.step} />
+            </Modal.Body>
+          </Modal>
+          {/*Loading*/}
+          <div className="loading-background">
+            <div className="loading">
+              <div className="lds-ripple">
+                <div></div>
+                <div></div>
               </div>
             </div>
           </div>
-        </Router>
-      )
-    }
+          <Footer />
+        </div>
+      </Router>
+    )
   }
 
 }
