@@ -6,7 +6,7 @@ class MyTooltip extends React.Component {
   render() {
     return (
       <OverlayTrigger
-        overlay={<Tooltip id="rand_Id">{this.props.tooltip}</Tooltip>} placement="top"
+        overlay={<Tooltip style={{whiteSpace: "pre-wrap"}} id="rand_Id">{this.props.tooltip}</Tooltip>} placement="top"
         delayShow={100} delayHide={150}
       >
         {this.props.children}
